refactor: reorganize websocket architecture
