+++ src-tauri/src/adaptor/controller/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod agent_session;
 pub(crate) mod agent_status_wiring;
 pub(crate) mod application_lifecycle;
 pub(crate) mod command;
+pub(crate) mod handler;
 pub(crate) mod notification_wiring;
 pub(crate) mod state;
 pub(crate) mod wiring;
