+++ src-tauri/src/adaptor/gateway/shared/mod.rs
@@ -1,1 +1,2 @@
 pub mod error_handling;
+pub mod ws_broadcaster;
