+++ src-tauri/src/adaptor/presenter/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod agent_status;
 pub(crate) mod workflow;
