+++ src-tauri/src/infrastructure/middleware/auth.rs
