+++ src-tauri/src/infrastructure/middleware/http_upgrade.rs
