+++ src-tauri/src/infrastructure/middleware/mod.rs
@@ -1,8 +1,7 @@
 mod auth;
-pub(crate) mod commands;
-mod http;
+mod http_upgrade;
 mod rate_limit;
-mod routing;
+pub(crate) mod server_control;
 mod session;
 
 use std::collections::HashMap;
@@ -10,10 +9,10 @@ use std::sync::Arc;
 
 use tokio::sync::Mutex;
 
+use crate::adaptor::gateway::shared::ws_broadcaster::WsBroadcaster;
 use crate::domain::app_config::ConfigRepository;
 use crate::usecase::agent_session::session::AgentStreamResyncReadModel;
 use crate::usecase::pty_session::query_service::PtySessionReplayReader;
-use crate::ws_bridge::WsBroadcaster;
 
 #[derive(Debug, Clone, serde::Serialize)]
 pub struct StartServerResult {
@@ -75,28 +74,5 @@ impl WsServerState {
     pub(crate) fn current_token(&self) -> Result<String, String> {
         let config = self.app_config.load().map_err(|e| e.to_string())?;
         Ok(config.server.token.clone())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::adaptor::protocol::deserialize_message;
-
-    #[test]
-    fn test_deserialize_invalid_json() {
-        let result = deserialize_message("not valid json at all");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_deserialize_empty_payload() {
-        let result = deserialize_message("");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_deserialize_missing_type_field() {
-        let result = deserialize_message(r#"{"data": "hello"}"#);
-        assert!(result.is_err());
     }
 }
