+++ src-tauri/src/infrastructure/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod agent_session;
 pub(crate) mod comment;
 pub(crate) mod file_watcher;
 pub mod git;
+pub(crate) mod middleware;
 pub(crate) mod platform;
 pub(crate) mod pty_session;
 pub(crate) mod telemetry;
